:recycle: Refactor: AuthService 리팩토링 - 단일 책임 원칙 적용 - 기존 AuthService.java 코드에 있던 로직 중, AuthService의 '인증'이라는 책임에 어긋나는 로직이 존재했음 ex) 사용자 생성, 사용자 기본 데이터 할당, 토큰 관리 등. 따라서, 해당 로직들을 단일책임 원칙에 맞게 전부 분리하였음 1. User 생성 및 기본 데이터 할당 로직 -> UserService 2. Refresh Token 관리 및 Redis 등록 등의 Token 관리 로직 -> TokenManagementService
그리고, 기존 RefreshTokenBackupRepository에 존재하던 N+1 문제를 야기할 수 있는 코드를 발견하여 해당 부분을 해결하였음
